alpha4
This is a bit debugged version. Some important bugs in changing the database "creatures.db" during each game are deleted. The version seems to work right(Every time I tested it all the 3 tables were filled correctly and the questions were chosen according to the number of zeroes in the answers in table "CORRELATIONS")